refactor: merge both operation modules
